// src/modules/informes/pages/rep-otros/pages/RepOtrosPage/types.ts
export interface TypesForm {
  tipo_reporte: string;
  bodegas: string[];
  fechas_rango: undefined;
  consecutivos: string[];
  productos: string[];
  estados: string[];
  documentos: string[];
  documentos_origen: string[];
  codigo_producto: string[];
  lote: string;
  nombre_producto: string;
  codigo_paciente: string;
  apellido1_paciente: string;
}